Fix MultipleSelect breaking when user presses backspace.
In the Select component it sends `null` back as the value when you press backspace, as if to say “clear the current value”, but we don’t have these semantics on a MultipleSelect. So just ignore the value that was passed if it doesn’t resolve to a real object from `this.props.values`.