Big update to bring compatibility with the latest changes in sixteen (especially around forms and the classes to use therein).
Removes a lot of things (e.g. help on hover/focus) due to the changes in sixteen.

This may well be broken in parts as it is not fully tested.

As it is possibly broken *but also* a large, breaking update to React UI, I have tagged it 1.0.0-alpha.1 in the hopes that 0.x versions don't get automatically updated, but also that it stays non-final. Wish me luck!

Closes #29.